+++ src/devices/src/virtio/vsock/muxer.rs
@@ -210,11 +210,9 @@ impl VsockMuxer {
             .epoll
             .ctl(ControlOperation::Delete, fd, &EpollEvent::default());
         if !evset.is_empty() {
-            let _ = self.epoll.ctl(
-                ControlOperation::Add,
-                fd,
-                &EpollEvent::new(evset, id as u64),
-            );
+            let _ = self
+                .epoll
+                .ctl(ControlOperation::Add, fd, &EpollEvent::new(evset, id));
         }
     }
